Fixed a bit on the intro Changed food and animal class to abstract Made the Game mechanic work ( input players, names of players, and amount of rounds.) Fixed the subclasses to the animals and aswell as the store menu, and the first part of buying animals. Started to work on the menu to buy the food. Made sell animals work. Working on the breeding option. Working on Breeding, starting with the methods. Made Breeding work. Starting on game mechanics such as switching player after a choice that has been made. Made the playerlist loop work, starting to work on game mechanics in store.
Made the method DecreaseDinoHealth work, decreasing the health of a players animals each round.

Working on fixing the problem with animals health reaching 0 and also more "cosmetic" problems such as text alignment and so on.

If a pets health reaches 0 it now dies. Also fixed 90% of the importance of kg of food in regards to health regen.
Probably done with the last touch tomorrow regarding the food.

Food now removed after it reaches 0kg owned.

Made decreaseDinoHealth work and also fixed on the playerloop.

Made endRound() , sellPlayerPets() and winner() methods work. The only thing left is a few bugfixes.

Added the method that removes a player from the game if certain conditions are met (No money AND no owned pets).
Fixed a try/catch method for the majority of the input scenarios.